requirements.txt
googlesearch-python==1.2.3
beautifulsoup4==4.12.2
selenium==4.14.0
openai==0.28.1
lxml==5.2.1
pydantic==2.5.2
sqlalchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0.post1
streamlit==1.28.2